move `UnaryNumber` to a separate file
